+++ lib/src/utils/data_converter.dart
@@ -88,6 +88,7 @@ class DataConverter {
         },
         pointFocusColor: theme?.interactionTheme.crosshairColor,
         pointSelectionColor: theme?.interactionTheme.selectionColor,
+        fontFamily: theme?.typographyTheme.fontFamily,
       );
     }).toList();
   }